refactor: update module paths in `plotting` and `processing` for consistency with `src` directory structure
